Moved opensearch modules
